[mem]: start using new alloc API - replace callsites

+++ include/dbg.h
@@ -5,5 +5,5 @@ void debug_print_registers();
 
 void debug_print_stack();
 
-void debug_print_memory(void *addr, uint64_t size);
+void debug_print_memory(void *addr, size_t size);
 #pragma once
